+++ src/lib/utils/useScreenDimensions.tsx
@@ -2,7 +2,7 @@ import { SafeAreaInsets } from "lib/types/SafeAreaInsets"
 import { createContext, useContext, useEffect, useState } from "react"
 import React from "react"
 import { Dimensions } from "react-native"
-import { SafeAreaProvider, useSafeAreaInsets } from "react-native-safe-area-context"
+import { useSafeAreaInsets } from "react-native-safe-area-context"
 
 export type ScreenOrientation = "landscape" | "portrait"
