+++ src/palette/elements/Text/__tests__/Text-tests.ts
@@ -1,5 +1,3 @@
-import { renderWithWrappersTL } from "lib/tests/renderWithWrappers"
-import React from "react"
 import { Text } from ".."
 import { endash, range } from "../../../helpers"
