--- src/palette/elements/Text/tokens.ts
@@ -1,119 +1,0 @@
-import {
-  TEXT_FONT_SIZES,
-  TEXT_LETTER_SPACING,
-  TEXT_LINE_HEIGHTS,
-  TEXT_VARIANT_NAMES as TEXT_TREATMENTS,
-  TEXT_VARIANTS as WEB_TEXT_VARIANTS,
-  TextTreatment as WebTextTreatment,
-} from "@artsy/palette-tokens/dist/typography/v2"
-import { ResponsiveValue, Theme, TLengthStyledSystem } from "styled-system"
-
-// get rid of these three
-export type TextFontSize = keyof typeof TEXT_FONT_SIZES
-type TextLetterSpacing = keyof typeof TEXT_LETTER_SPACING
-export type TextLineHeight = keyof typeof TEXT_LINE_HEIGHTS
-
-/**
- * font-families
- */
-export const TEXT_FONTS_V2 = {
-  sans: "Unica77LL-Regular",
-  serif: "ReactNativeAGaramondPro-Regular",
-}
-export const TEXT_FONTS_V3 = {
-  sans: {
-    regular: "Unica77LL-Regular",
-    italic: "Unica77LL-Italic",
-    medium: "Unica77LL-Medium",
-    mediumItalic: "Unica77LL-MediumItalic",
-  },
-}
-export const TEXT_FONTS = TEXT_FONTS_V2
-
-/**
- * em-units don't exist on React Native so we convert it to a number
- * which will be evaluated as px based on the given font-size.
- */
-export const calculateLetterSpacing = (fontSize: TextFontSize, letterSpacing: TextLetterSpacing): number => {
-  const tracking = parseFloat(TEXT_LETTER_SPACING[letterSpacing])
-  const size = parseInt(TEXT_FONT_SIZES[fontSize], 10)
-  return size * tracking
-}
-
-/**
- * unitless line-heights don't exist on React Native so we convert it
- * to a px string. Since unitless line-heights are valid/normal, styled-system
- * won't convert it to px.
- */
-export const calculateLineHeight = (fontSize: TextFontSize, lineHeight: TextLineHeight): number => {
-  const size = parseInt(TEXT_FONT_SIZES[fontSize], 10)
-  return size * TEXT_LINE_HEIGHTS[lineHeight]
-}
-
-export interface TextTreatment extends Omit<WebTextTreatment, "lineHeight" | "letterSpacing"> {
-  lineHeight: number
-  letterSpacing?: string
-}
-
-/**
- * Utilizes only the small treatments on mobile
- */
-export const TREATMENTS = WEB_TEXT_VARIANTS.small
-
-/**
- * TextTreatments
- */
-export type TextTreatments = {
-  [K in typeof TEXT_TREATMENTS[number]]: TextTreatment
-}
-
-export type TextVariant = keyof TextTreatments
-
-/**
- * iOS-specific typographic treatments
- */
-export const TEXT_VARIANTS = (Object.keys(TREATMENTS) as Array<keyof typeof TREATMENTS>).reduce((acc, name) => {
-  const { fontSize, fontWeight, letterSpacing, lineHeight } = TREATMENTS[name]
-
-  return {
-    ...acc,
-    [name]: {
-      fontSize,
-      fontWeight,
-
-      // Convert `letterSpacing`
-      ...(letterSpacing
-        ? {
-            letterSpacing: calculateLetterSpacing(fontSize as TextFontSize, letterSpacing as TextLetterSpacing),
-          }
-        : {}),
-
-      // Convert `lineHeight`
-      ...(lineHeight
-        ? {
-            lineHeight: calculateLineHeight(fontSize as TextFontSize, lineHeight as TextLineHeight),
-          }
-        : {}),
-    },
-  }
-}, {} as TextTreatments)
-
-type ThemedKey = ResponsiveValue<string | number | symbol, Required<Theme<TLengthStyledSystem>>>
-
-/**
- * Type-guard that determines whether or not value is within theme
- */
-export const isControlledLetterSpacing = (key: ThemedKey): key is TextLetterSpacing =>
-  Object.keys(TEXT_LETTER_SPACING).includes(String(key))
-
-/**
- * Type-guard that determines whether or not value is within theme
- */
-export const isControlledLineHeight = (key: ThemedKey): key is TextLineHeight =>
-  Object.keys(TEXT_LINE_HEIGHTS).includes(String(key))
-
-/**
- * Type-guard that determines whether or not value is within theme
- */
-export const isControlledFontSize = (key: ThemedKey): key is TextFontSize =>
-  Object.keys(TEXT_FONT_SIZES).includes(String(key))
